+++ tests/RedirectMiddlewareTest.php
@@ -275,7 +275,6 @@ class RedirectMiddlewareTest extends TestCase
     /**
      * @testWith ["digest"]
      *           ["ntlm"]
-     *           ["bearer"]
      */
     public function testRemoveCurlAuthorizationOptionsOnRedirectCrossHost($auth)
     {
@@ -294,6 +293,30 @@ class RedirectMiddlewareTest extends TestCase
                     isset($options['curl'][\CURLOPT_USERPWD]),
                     'curl options still contain CURLOPT_USERPWD entry'
                 );
+
+                return new Response(200);
+            },
+        ]);
+        $handler = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handler]);
+        $client->get('http://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+    }
+
+    public function testRemoveCurlBarerAuthorizationOptionsOnRedirectCrossHost()
+    {
+        if (PHP_VERSION_ID < 70300) {
+            $this->markTestSkipped('This test requires PHP version 7.3 or higher.');
+        }
+        if (!defined('\CURLOPT_HTTPAUTH')) {
+            self::markTestSkipped('ext-curl is required for this test');
+        }
+        if (!defined('\CURLAUTH_BEARER')) {
+            self::markTestSkipped('curl version 7.61.0 or higher is required for this test');
+        }
+
+        $mock = new MockHandler([
+            new Response(302, ['Location' => 'http://test.com']),
+            static function (RequestInterface $request, $options) {
                 self::assertFalse(
                     isset($options['curl'][\CURLOPT_XOAUTH2_BEARER]),
                     'curl options still contain CURLOPT_XOAUTH2_BEARER entry'
@@ -304,13 +327,12 @@ class RedirectMiddlewareTest extends TestCase
         ]);
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
-        $client->get('http://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+        $client->get('http://example.com?a=b', ['auth' => ['testtoken', '', 'bearer']]);
     }
 
     /**
      * @testWith ["digest"]
      *           ["ntlm"]
-     *           ["bearer"]
      */
     public function testRemoveCurlAuthorizationOptionsOnRedirectCrossPort($auth)
     {
@@ -329,6 +351,30 @@ class RedirectMiddlewareTest extends TestCase
                     isset($options['curl'][\CURLOPT_USERPWD]),
                     'curl options still contain CURLOPT_USERPWD entry'
                 );
+
+                return new Response(200);
+            },
+        ]);
+        $handler = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handler]);
+        $client->get('http://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+    }
+
+    public function testRemoveCurlBearerAuthorizationOptionsOnRedirectCrossPort()
+    {
+        if (PHP_VERSION_ID < 70300) {
+            $this->markTestSkipped('This test requires PHP version 7.3 or higher.');
+        }
+        if (!defined('\CURLOPT_HTTPAUTH')) {
+            self::markTestSkipped('ext-curl is required for this test');
+        }
+        if (!defined('\CURLAUTH_BEARER')) {
+            self::markTestSkipped('curl version 7.61.0 or higher is required for this test');
+        }
+
+        $mock = new MockHandler([
+            new Response(302, ['Location' => 'http://example.com:81/']),
+            static function (RequestInterface $request, $options) {
                 self::assertFalse(
                     isset($options['curl'][\CURLOPT_XOAUTH2_BEARER]),
                     'curl options still contain CURLOPT_XOAUTH2_BEARER entry'
@@ -339,13 +385,12 @@ class RedirectMiddlewareTest extends TestCase
         ]);
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
-        $client->get('http://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+        $client->get('http://example.com?a=b', ['auth' => ['testtoken', '', 'bearer']]);
     }
 
     /**
      * @testWith ["digest"]
      *           ["ntlm"]
-     *           ["bearer"]
      */
     public function testRemoveCurlAuthorizationOptionsOnRedirectCrossScheme($auth)
     {
@@ -364,6 +409,30 @@ class RedirectMiddlewareTest extends TestCase
                     isset($options['curl'][\CURLOPT_USERPWD]),
                     'curl options still contain CURLOPT_USERPWD entry'
                 );
+
+                return new Response(200);
+            },
+        ]);
+        $handler = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handler]);
+        $client->get('https://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+    }
+
+    public function testRemoveCurlBearerAuthorizationOptionsOnRedirectCrossScheme()
+    {
+        if (PHP_VERSION_ID < 70300) {
+            $this->markTestSkipped('This test requires PHP version 7.3 or higher.');
+        }
+        if (!defined('\CURLOPT_HTTPAUTH')) {
+            self::markTestSkipped('ext-curl is required for this test');
+        }
+        if (!defined('\CURLAUTH_BEARER')) {
+            self::markTestSkipped('curl version 7.61.0 or higher is required for this test');
+        }
+
+        $mock = new MockHandler([
+            new Response(302, ['Location' => 'http://example.com?a=b']),
+            static function (RequestInterface $request, $options) {
                 self::assertFalse(
                     isset($options['curl'][\CURLOPT_XOAUTH2_BEARER]),
                     'curl options still contain CURLOPT_XOAUTH2_BEARER entry'
@@ -374,13 +443,12 @@ class RedirectMiddlewareTest extends TestCase
         ]);
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
-        $client->get('https://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+        $client->get('https://example.com?a=b', ['auth' => ['testtoken', '', 'bearer']]);
     }
 
     /**
      * @testWith ["digest"]
      *           ["ntlm"]
-     *           ["bearer"]
      */
     public function testRemoveCurlAuthorizationOptionsOnRedirectCrossSchemeSamePort($auth)
     {
@@ -410,6 +478,34 @@ class RedirectMiddlewareTest extends TestCase
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
         $client->get('https://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
+    }
+
+    public function testRemoveCurlBearerAuthorizationOptionsOnRedirectCrossSchemeSamePort()
+    {
+        if (PHP_VERSION_ID < 70300) {
+            $this->markTestSkipped('This test requires PHP version 7.3 or higher.');
+        }
+        if (!defined('\CURLOPT_HTTPAUTH')) {
+            self::markTestSkipped('ext-curl is required for this test');
+        }
+        if (!defined('\CURLAUTH_BEARER')) {
+            self::markTestSkipped('curl version 7.61.0 or higher is required for this test');
+        }
+
+        $mock = new MockHandler([
+            new Response(302, ['Location' => 'http://example.com:80?a=b']),
+            static function (RequestInterface $request, $options) {
+                self::assertFalse(
+                    isset($options['curl'][\CURLOPT_XOAUTH2_BEARER]),
+                    'curl options still contain CURLOPT_XOAUTH2_BEARER entry'
+                );
+
+                return new Response(200);
+            },
+        ]);
+        $handler = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handler]);
+        $client->get('https://example.com?a=b', ['auth' => ['testtoken', '', 'bearer']]);
     }
 
     /**
@@ -442,14 +538,16 @@ class RedirectMiddlewareTest extends TestCase
         $client->get('http://example.com?a=b', ['auth' => ['testuser', 'testpass', $auth]]);
     }
 
-    /**
-     * @testWith ["bearer"]
-     *
-     */
-    public function testNotRemoveCurlBearerAuthorizationOptionsOnRedirect($auth)
+    public function testNotRemoveCurlBearerAuthorizationOptionsOnRedirect()
     {
-        if (!defined('\CURLOPT_HTTPAUTH') || !defined('\CURLOPT_XOAUTH2_BEARER')) {
+        if (PHP_VERSION_ID < 70300) {
+            $this->markTestSkipped('This test requires PHP version 7.3 or higher.');
+        }
+        if (!defined('\CURLOPT_HTTPAUTH')) {
             self::markTestSkipped('ext-curl is required for this test');
+        }
+        if (!defined('\CURLAUTH_BEARER')) {
+            self::markTestSkipped('curl version 7.61.0 or higher is required for this test');
         }
 
         $mock = new MockHandler([
@@ -465,7 +563,7 @@ class RedirectMiddlewareTest extends TestCase
         ]);
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
-        $client->get('http://example.com?a=b', ['auth' => ['testtoken', '', $auth]]);
+        $client->get('http://example.com?a=b', ['auth' => ['testtoken', '', 'bearer']]);
     }
 
     public static function crossOriginRedirectProvider()